refactor(dictation): 抽出 dictation_session 子模块
把会话生命周期函数从 coordinator/dictation.rs 机械式搬到同级
coordinator/dictation_session.rs（行为保持）：request_stop_during_starting、
begin_session、start_recorder_for_starting、spawn_recorder_error_monitor、
abort_recording_with_error、start_recorder_and_enter_listening、
finish_starting_session、dictation_error_code、cancel_session、
eligible_polish_context_turns。新模块 use super::resources::* 取资源辅助函数。
coordinator.rs glob 重导出并把 begin_session/cancel_session/
request_stop_during_starting/dictation_error_code 的命名导入改到
dictation_session::；tests.rs 的 abort_recording_with_error 路径同步；
dictation.rs 去掉随之失用的 request_stop_during_starting_state 导入，
测试 eligible_polish_context_turns 改走 crate::coordinator:: 路径。
